+++ src/components/QuizForm/styles.js
@@ -1,0 +1,32 @@
+
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    marginLeft:10,
+    marginRight: 10,
+    alignItems: 'stretch',
+    justifyContent: 'center',
+  },
+  btn: {
+    backgroundColor: '#E53224',
+    padding: 10,
+    paddingLeft: 50,
+    paddingRight: 50,
+    justifyContent : 'center',
+    alignItems: 'center',
+    borderRadius: 5
+  },
+  btnText: {
+    color: '#fff'
+  },
+  input: {
+    width: 200,
+    height: 44,
+    padding: 8,
+    borderWidth: 1,
+    borderColor: '#757575',
+  }
+});