+++ src/constants/index.js
@@ -1,6 +1,6 @@
 
 // base quiz data from the project description
-const QuizData = {
+export const quizData = {
   React: {
     title: 'React',
     questions: [
@@ -24,3 +24,17 @@ const QuizData = {
     ]
   }
 }
+
+export const newTest = {React2: {
+      title: 'React2',
+      questions: [
+        {
+          question: 'What is React?',
+          answer: 'A library for managing user interfaces'
+        },
+        {
+          question: 'Where do you make Ajax requests in React?',
+          answer: 'The componentDidMount lifecycle event'
+        }
+      ]
+}}