+++ src/views/AddQuiz.js
@@ -1,14 +1,21 @@
 import React, { Component } from 'react'
 import { View, Text } from 'react-native'
+import { connect } from 'react-redux'
 
-export default class AddQuiz extends Component {
+import QuizForm from '../components/QuizForm'
+
+class AddQuiz extends Component {
 	render(){
 		return (
-
-			<View>
-				<Text> Add Quiz view </Text>
-			</View>
-
+			<QuizForm />
 		)
 	}
 }
+
+function mapStateToProps(state){
+
+	// no state to pass yet, but will pass all for now to complete redux connection
+	return {state}
+}
+
+export default connect(mapStateToProps)(AddQuiz)