+++ CHANGELOG.md
@@ -1,13 +1,22 @@
 # CHANGELOG - Mobile Flashcards - react-native && react-redux  
 
-## v0.2
+## v0.3
 	- TODO:
-		+ create static data
+		+ create base components including:
+			- QuizList
+			- QuizItem
+			- AddQuiz
+			- Quiz
+		+ create base navigation for quiz list and quiz components
+	- COMPLETED:
+		+ hook up redux
 
+## v0.2
 	- COMPLETED:
 		+ create base components including:
 			- navigation
 			- base views
+		+ create static data
 
 ## v0.1 
 	COMPLETED: